add note per user functionallity

+++ server/server.js
@@ -31,6 +31,23 @@ const corsOptions = {
   optionsSuccessStatus: 200,
 };
 
+function authenticateToken(req, res, next){
+  const authHeader = req.headers['authorization'];
+  const token = authHeader && authHeader.split(' ')[1];
+
+  if (!token) {
+    return res.sendStatus(401).json({message : "No token found"})
+  }
+
+  jwt.verify(token, "jwt-secret", (err, user) => {
+    if (err) {
+      return res.sendStatus(403).json({message:"Invalid token"})
+    }
+    req.user = user;
+    next();
+  });
+};
+
 app.use(cors(corsOptions));
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
@@ -48,18 +65,19 @@ app.use(express.json());
 app.post("/register", async (req, res) => {
   const {firstName, lastName, email, password} = req.body;
   try {
-    const result = await db.query('SELECT * FROM users WHERE "email" = $1', [email]);
+    const result = await db.query('SELECT * FROM users WHERE "email" = $1', [email.toLowerCase()]);
     if (result.rows.length > 0) {
-      return res.status(400).json({ message: "Email already registered."});
+      const user = result.rows[0]
+      return res.status(400).json({ message: "Email already registered.", user: user});
     }
 
     const passwordHash = await bcrypt.hash(password, 10);
 
-    await db.query('INSERT INTO users ("email", "firstName", "lastName", "password") VALUES ($1, $2, $3, $4)', [email, firstName, lastName, passwordHash]);
-    
-    const token = jwt.sign({email}, "jwt-secret", {expiresIn: "1h"});
+    result = await db.query('INSERT INTO users ("email", "firstName", "lastName", "password") VALUES ($1, $2, $3, $4)', [email.toLowerCase(), firstName, lastName, passwordHash]);
+    const user = result.rows[0]
+    const token = jwt.sign({ id: user.id, email: user.email}, "jwt-secret", {expiresIn: "1h"});
 
-    res.status(201).json({message: "user registered", token});
+    res.status(201).json({message: "user registered", token, user});
     // Here i should login and authenticate the new user
     
   } catch (error) {
@@ -72,7 +90,7 @@ app.post("/register", async (req, res) => {
 app.post("/login", async (req, res) => {
   const { email, password } = req.body;
   try {
-    const result = await db.query('SELECT * FROM users WHERE "email" = $1', [email]);
+    const result = await db.query('SELECT * FROM users WHERE "email" = $1', [email.toLowerCase()]);
     if (result.rows.length === 0) {
       res.status(404).json({message: "User not found"});
     }
@@ -83,8 +101,8 @@ app.post("/login", async (req, res) => {
       return res.status(401).json({message: "Invalid password."});
     }
 
-    const token = jwt.sign({ email: user.email }, "jwt-secret", { expiresIn: "1h" });
-    res.json({message: "Login successful", token});
+    const token = jwt.sign({ id: user.id, email: user.email }, "jwt-secret", { expiresIn: "1h" });
+    res.json({message: "Login successful", token, user: user});
     
   } catch (error) {
     console.error("Error fetching data", error);
@@ -94,9 +112,9 @@ app.post("/login", async (req, res) => {
 
 
 // API FETCH notes FROM db
-app.get("/", async (req, res) => {
+app.get("/", authenticateToken ,async (req, res) => {
   try {
-    const result = await db.query("SELECT * FROM notes");
+    const result = await db.query("SELECT * FROM notes WHERE user_id = $1", [req.user.id]);
     if (result.rows.length > 0) {
       const notes = result.rows;
       res.json(notes);
@@ -109,12 +127,19 @@ app.get("/", async (req, res) => {
   }
 });
 
+// API GETUSER
+// app.get('/user', authenticateToken, async(req, res) => {
+//   console.log(req.user.id)
+//   const response = await db.query("SELECT * FROM users WHERE id = $1", [req.user.id])
+//   console.log(response);
+// })
+
 // API INSERT notes INTO db
-app.post("/add", async (req, res) => {
+app.post("/add", authenticateToken, async (req, res) => {
     const { note } = req.body
     try {
       if (note.title != "" && note.content != ""){
-        const result = await db.query("INSERT INTO notes (title, content) VALUES ($1, $2) RETURNING *", [note.title, note.content]);
+        const result = await db.query("INSERT INTO notes (title, content, user_id) VALUES ($1, $2, $3) RETURNING *", [note.title, note.content, req.user.id]);
         res.status(200).json(result.rows[0]);
       }
     } catch (error) {
@@ -124,10 +149,10 @@ app.post("/add", async (req, res) => {
 });
 
 // API DELETE notes FROM db
-app.post("/delete", async (req, res) => {
+app.post("/delete", authenticateToken, async (req, res) => {
   const { id } = req.body;
   try {
-    const result = await db.query("DELETE FROM notes WHERE id = $1 RETURNING *", [id]);
+    const result = await db.query("DELETE FROM notes WHERE id = $1 AND user_id = $2 RETURNING *", [id, req.user.id]);
     res.status(200).json(result.rows[0]);
   } catch (error){
     console.error("Error deleting note ", error);
